+++ CHANGES.rst
@@ -62,6 +62,8 @@ Improvements
   and/or all its rooms (:pr:`6566`, thanks :user:`SegiNyn`)
 - Support alternative names in predefined affiliations and make its search more powerful
   (:pr:`6758`)
+- Add setting to disallow entering custom affiliations when predefined affiliations are used
+  (:pr:`6809`)
 
 Bugfixes
 ^^^^^^^^